requirements.txt
pynput
mss
opencv-python
pyserial
numpy